refactor: move framework-tax microbenchmarks to benchmarks/micro/
Relocates the four-stack framework-tax matrix from internal/contractspike
to benchmarks/micro/{nethttp,gin,huma,gombit}, one package per row,
sharing resource types, Huma route registration, and correctness
assertions via benchmarks/micro/scenario.

This matches the benchmarks/ tree issue #141 itself suggests (a
sibling micro/ next to apps/, workloads/, etc.) instead of the
internal/contractspike placement from the previous commit, which read
the issue's "reuse the M0 spike, don't create an unrelated
implementation" instruction as "stay in the same directory." In
practice the new matrix never touched the spike's existing code
(server.go/server_test.go, the widget routes, the committed
openapi.json) — it only shared a directory with it. The spike's result
and intent are preserved by cross-linking
(docs/spikes/M0-2_HUMA_GIN_SPIKE.md -> benchmarks/README.md), which
doesn't require physical colocation.

One package per row also gets the contract.Install process-isolation
requirement (constructing a framework.App mutates Huma's process-global
huma.NewError) for free from go test's one-process-per-package model,
for all four rows uniformly, instead of a special-cased boundary for
just the Gombit row.

internal/contractspike is back to its original M0-2 spike state
(server.go, server_test.go only); its OpenAPI-drift test is
unaffected since it no longer shares a package with anything that
calls contract.Install.

Updates docs/plans/BENCH-1-benchmark-suite.md's Phase 1/2 status and
repo-structure section, docs/GOMBIT_BUILD_PLAN.md's BENCH-1 entry
wording, and the M0 spike doc's cross-link accordingly.
