+++ docs/GOMBIT_BUILD_PLAN.md
@@ -231,7 +231,7 @@ Each is a **future epic**, explicitly out of v0.1. Do not create these as active
 
 ### BENCH — Reproducible benchmarks (POST-v0.1)
 
-- **[BENCH-1] Reproducible framework benchmark suite + README performance results** — checked-in `benchmarks/` harness quantifying Gombit's abstraction cost (net/http → Gin → Huma+Gin → Gombit) and realistic PostgreSQL CRUD/auth performance against Gin+GORM (primary control) and Django+DRF/Rails/Laravel/NestJS (ecosystem context), plus a generated README `## Performance` section. Extends the M0-2 Huma/Gin spike benchmark (`internal/contractspike`) rather than replacing it. AC: see issue #141 §25. Implementation plan: [plans/BENCH-1-benchmark-suite.md](plans/BENCH-1-benchmark-suite.md). deps: none (post-v0.1, no locked architecture change). size: XL. labels: `infra`, `devx`, `tests`, `ci`.
+- **[BENCH-1] Reproducible framework benchmark suite + README performance results** — checked-in `benchmarks/` harness quantifying Gombit's abstraction cost (net/http → Gin → Huma+Gin → Gombit) and realistic PostgreSQL CRUD/auth performance against Gin+GORM (primary control) and Django+DRF/Rails/Laravel/NestJS (ecosystem context), plus a generated README `## Performance` section. Supersedes the M0-2 Huma/Gin spike benchmark (`internal/contractspike`) as the ongoing framework-tax measurement while preserving it untouched as historical record. AC: see issue #141 §25. Implementation plan: [plans/BENCH-1-benchmark-suite.md](plans/BENCH-1-benchmark-suite.md). deps: none (post-v0.1, no locked architecture change). size: XL. labels: `infra`, `devx`, `tests`, `ci`.
 
 ---
